+++ scripts/bootstrap_proxy.sh
@@ -1,7 +1,17 @@
 #! /bin/bash
 
-echo -e "\x1b[33mChecking for existing tinyweather-proxy service file...\x1b[0m"
+# Check if both arguments are provided
+if [ $# -ne 2 ]; then
+    echo -e "\x1b[31mError: Please provide both address and port arguments.\x1b[0m"
+    echo -e "\x1b[33mUsage: $0 <address> <port>\x1b[0m"
+    exit 1
+fi
 
+# Store command line arguments
+ADDRESS=$1
+PORT=$2
+
+echo -e "\x1b[33mChecking for existing tinyweather-proxy service file...\x1b[0m"
 if [[ -f /etc/systemd/system/tinyweather-proxy.service ]]; then 
     systemctl stop tinyweather-proxy.service
     systemctl disable tinyweather-proxy.service
@@ -15,44 +25,33 @@ fi
 echo -e "\x1b[32mBeginning the bootstrapping of tinyweather-proxy\x1b[0m"
 echo -e "\x1b[33mBuilding tinyweather-proxy with 'zig build'...\x1b[0m"
 zig build 
-
-
 echo -e "\x1b[33mPreparing /opt/tinyweather directory...\x1b[0m"
 mkdir -p /opt/tinyweather
-
 echo -e "\x1b[33mRemoving existing tinyweather-proxy binary from /opt/tinyweather (if it exists)...\x1b[0m"
 rm -f /opt/tinyweather/tinyweather-proxy
-
 echo -e "\x1b[33mMoving newly built tinyweather-proxy executable to /opt/tinyweather\x1b[0m"
 mv ./zig-out/bin/tinyweather-proxy /opt/tinyweather
-
 echo -e "\x1b[33mCreating systemd service file at /etc/systemd/system/tinyweather-proxy.service...\x1b[0m"
 touch /etc/systemd/system/tinyweather-proxy.service
-
 echo -e "\x1b[33mWriting service configuration to /etc/systemd/system/tinyweather-proxy.service...\x1b[0m"
 echo "
 [Unit]
 Description=Tinyweather proxy Service
 After=network.target
-
 [Service]
 Type=simple
 Restart=always
 RestartSec=5s
-ExecStart=/opt/tinyweather/tinyweather-proxy --address 10.0.2.14 --port 8081 
+ExecStart=/opt/tinyweather/tinyweather-proxy --address $ADDRESS --port $PORT 
 WorkingDirectory=/opt/tinyweather
-
 [Install]
 WantedBy=multi-user.target" >> /etc/systemd/system/tinyweather-proxy.service
 echo -e "\x1b[32mService configuration written to /etc/systemd/system/tinyweather-proxy.service\x1b[0m"
-
 echo -e "\x1b[33mReloading systemd daemon to recognize the new tinyweather-proxy service...\x1b[0m"
 systemctl daemon-reload
 echo -e "\x1b[32mSystemd daemon reloaded successfully.\x1b[0m"
-
 echo -e "\x1b[33mStarting the tinyweather-proxy service...\x1b[0m"
 systemctl start tinyweather-proxy
 systemctl enable tinyweather-proxy
 echo -e "\x1b[32mtinyweather-proxy service started.\x1b[0m"
-
 systemctl status tinyweather-proxy
